requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.23
PyMySQL==1.1.0
python-dotenv==1.0.0
Flask-Migrate==4.0.5
gunicorn==21.2.0
Flask-JWT-Extended==4.5.3
bcrypt==4.0.1
Werkzeug==2.3.7
Flask-Limiter==3.5.0
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Flask-WTF